+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Added
+* Added `.overrideToolchain` as a convenience for using a custom rust toolchain
+
 ## [0.5.0] - 2022-06-12
 
 ### Changed